fastapi-sample/requirements.txt
openai
python-dotenv
langchain
langchain_openai
langchain_community
uvicorn
fastapi
websockets
